firmware: allow running spi at max speed (18 Mbps)
when running the stm32f0 using HSI48, the spi peripheral can't reach the max
speed and is limited to 12 Mbps. By reducing the main clock to 36 MHz using a
PLL, it is possible to increase the spi speed.

according to my experiments, it makes no difference in practice, but I'll
leave the code in, protected by a macro.